Resolve seed script regression
