+++ src/db/seed/seedScript.js
@@ -1,0 +1,7 @@
+import { seedDatabase } from "./seedScript";
+
+async function runSeed() {
+	await seedDatabase();
+}
+
+runSeed();
